requirements.txt
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
dominate==2.6.0
Flask-Script==2.0.6
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint-plugin-utils==0.6
python-dotenv==0.17.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
